pyproject.toml
[project]
name = "health-bot"
version = "0.1.0"
description = "Telegram bot for tracking daily health habits"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-telegram-bot[job-queue]>=20.7",
    "google-api-python-client>=2.111.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.0",
    "python-dotenv>=1.0.0",
    "pytz>=2024.1",
]

[project.scripts]
health-bot = "bot.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["bot"]
